requirements.txt
asgiref==3.8.1
Django==4.2.16
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.0
sqlparse==0.5.1
typing_extensions==4.12.2
